Removed ArchiveActivity.java and actually added ArchiveActivityNew.java and the new layout xml file.
