+++ res/layout/archive_item_layout.xml
@@ -1,0 +1,11 @@
+<?xml version="1.0" encoding="utf-8"?>
+<LinearLayout
+  xmlns:android="http://schemas.android.com/apk/res/android"
+  android:orientation="horizontal"
+  android:layout_width="match_parent"
+  android:layout_height="match_parent">
+    <CheckBox android:layout_height="wrap_content" android:id="@+id/archive_item_bookmarked" android:layout_width="wrap_content" style="?android:attr/starStyle"></CheckBox>
+	<TextView android:text="# - Comic Title" android:layout_width="wrap_content"
+		android:layout_height="wrap_content" android:id="@+id/archive_item_textView"
+		android:textSize="30px"></TextView>
+</LinearLayout>
